+++ HISTORY.rst
@@ -7,6 +7,7 @@ Development
 +++++++++++
 - Template tag `bootstrap_field` now allows 3 values for `show_label`: `True`, `False` / `'sr-only'` and `'skip'`. In the case of `False` / `'sr-only'` the label is hidden but present for screen readers. When `show_label` is set to `'skip'`, the label is not generated at all.
 - Fix validation on input groups (#122)
+- Apply form-check, form-check-label and form-check-input classes to appropriate tags for RadioSelect and CheckboxSelectMultiple widgets
 
 0.0.8 (2019-03-12)
 ++++++++++++++++++